==> jupyter_notebooks_analysis/notebooks/6_compute_nb_data.ipynb <==
# Compute Notebook Data

This notebook is devoted to calculating notebok statistics for further analysis. There are generally three types of stats we aim to calculate:
1. Notebook metadata stored in the notebook itself (e.g., nbformat, kernel_lang)
2. Descriptive information about cells, their type, and length
3. Descriptive information about the content of cells (import statements, keywords, links)

import os
import re
import time
import json
import datetime

import numpy as np
import pandas as pd
import seaborn as sns

%matplotlib inline
%load_ext line_profiler

## 0. Load notebook, repo, and readme metadata

# Notebook metadata
df_nb = pd.read_csv('../data/csv/nb_metadata.csv')
df_nb.rename(columns = {'Unnamed: 0':'nb_id'}, inplace = True)
print(df_nb.shape)
df_nb.head()

# Repository metadata
df_repo = pd.read_csv('../data/csv/repo_metadata.csv')
del df_repo['Unnamed: 0']
print(df_repo.shape)
df_repo.head()

# Readme metadtata
df_readme = pd.read_csv('../data/csv/repo_readme.csv')
del df_readme['Unnamed: 0']
print(df_readme.shape)
df_readme.head()

## 1. Scrape Additional nb data from nb file

We currently only have the metadata provided from Github's API. Now we can look at the nb files themselves and get the metadata from within the notebook (nbformat, kernel, langauge) and compute stats about each notebook (type and num of cell, and so on).

For now let's test this metadata extraction with a subset of 1000 nbs. And if that works we can then apply the extraction code to all notebooks.

df_nb_sample = df_nb.sample(1000)
df_nb_sample.shape

def write_to_log(msg):
    f = '../logs/nb_parse_log.txt'
    log_file = open(f, "a")
    log_file.write(msg + "\n")
    
def get_nb_metadata(df):
    # create blank columns for the data
    df["nbformat"] = None
    df["nbformat_minor"] = None
    df["kernel_lang"] = None
    df["kernel_name"] = None
    df["lang_name"] = None
    df["lang_version"] = None
    df['num_cells'] = None
    
    # track time to process files
    count = 0    
    time1 = time.time()
    
    for index, row in df.iterrows():
        # track time to process files
        count += 1
        if count % 10000 == 0:
            print('%s rows in %s seconds' % (count, time.time() - time1))

        f = '../data/notebooks/nb_%s.ipynb' % row['nb_id']
        with open(f) as data_file:
            date_string = datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S")

            try:
                data = json.load(data_file)
            except:
                # track files we could not open
                msg = '%s: had trouble loading %s' % (date_string, row['nb_id'])
                write_to_log(msg)
                continue

            if isinstance(data, dict): 
                keys = data.keys()
            else:
                keys = []

            # get nb top level metadata
            if 'nbformat' in keys:
                df.set_value(index, 'nbformat', data['nbformat'])
            if 'nbformat_minor' in keys:
                df.set_value(index, 'nbformat_minor', data['nbformat_minor'])
            
            # get information about the number of cells
            if 'cells' in keys:
                df.set_value(index, 'num_cells', len(data['cells']))
            elif 'worksheets' in keys:
                num_cells = 0
                for w in data['worksheets']:
                    num_cells += len(w['cells'])
                df.set_value(index, 'num_cells', num_cells)

            # get info from the metadata dictionary
            if 'metadata' in keys:
                metadata_keys = data['metadata'].keys()

                if 'kernelspec' in metadata_keys:
                    kernel_keys = data['metadata']['kernelspec'].keys()
                    if 'language' in kernel_keys:
                        df.set_value(index, 'kernel_lang', data['metadata']['kernelspec']['language'])
                    if 'display_name' in kernel_keys:
                        df.set_value(index, 'kernel_name', data['metadata']['kernelspec']['display_name'])

                if 'language_info' in metadata_keys:
                    lang_keys = data['metadata']['language_info'].keys()
                    if 'name' in lang_keys:
                        df.set_value(index, 'lang_name', data['metadata']['language_info']['name'])
                    if 'version' in lang_keys:
                        df.set_value(index, 'lang_version', data['metadata']['language_info']['version'])

                elif 'language' in metadata_keys:
                    df.set_value(index, 'lang_name', data['metadata']['language'])

We eventually got the code to work on the 1000 sample notebooks, so now we will apply it to the whole dataset. Here, as elsewhere I track how long the extraction is taking.

# get metadata and save updated dataframe to a csv file
get_nb_metadata(df_nb)
df_nb.to_csv('../data/csv/nb_metadata_w_nb.csv')

## 2. Get Data about Cell Types

I want to track data on each cell. I think the best way to do this is to have a dataframe with a separate row for each cell. I think I want to track the following data. Some of the difficulty in writing this extraction code will be handling nbformat 3.0 that use a different format to store notebook data.

1. nb_id
2. workbook_index
2. cell_index
3. cell_type
3. num_words (raw, markdown and headings)
4. lines of code (code cells)
5. num_execute_result
6. execute_result_keys
7. num_error
8. error_names
9. error_values
10. num_stream
11. num_display_data
12. display_data_keys

def write_to_log(msg):
    f = '../logs/cell_parse_log.txt'
    log_file = open(f, "a")
    log_file.write(msg + "\n")
    #log_file.close()   
    
def get_all_cell_data(df, tracking_trigger = 10000):
    
    all_cells = []      
    count = 0    
    time1 = time.time()
    
    for index, row in df.iterrows():
        
        # track progress through notebooks
        count += 1
        if count % tracking_trigger == 0:
            print('%s nbs in %s seconds' % (count, time.time() - time1))

        f = '../data/notebooks/nb_%s.ipynb' % row['nb_id']
        with open(f) as nb_file:
            date_string = datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S")

            try:
                data = json.load(nb_file)
            except:
                msg = '%s: had trouble loading nb %s' % (date_string, row['nb_id'])
                write_to_log(msg)
                continue

            if isinstance(data, dict): 
                keys = data.keys()
            else:
                keys = []

            # get data for each cell, v 4.0
            if 'cells' in keys:
                for i, c in enumerate(data['cells']):
                    cell_data = get_cell_data(c, i, row['nb_id'])
                    all_cells.append(cell_data)
            
            # get data for each cell, v 3.0
            elif 'worksheets' in keys:
                for j, w in enumerate(data['worksheets']):
                    
                    if isinstance(w, dict): 
                        worksheet_keys = w.keys()
                    else:
                        keys = []
                    
                    if 'cells' in worksheet_keys:
                        for k, c in enumerate(w['cells']):
                            cell_data = get_cell_data(c, k, row['nb_id'], j)
                            all_cells.append(cell_data)
                
    return all_cells
    
def get_cell_data(cell, index, nb_id, worksheet_index = None):
    nbformat_3_mimes = ['text', 'latex', 'png', 'jpeg', 'svg', 'html', 'javascript', 'json', 'pdf', 'metadata']
    
    if isinstance(cell, dict): 
        cell_keys = cell.keys()
    else:
        cell_keys = [] 
    
    if 'cell_type' in cell_keys:
        cell_type = cell['cell_type']
    else:
        cell_type = None
    
    
    if cell_type in ['raw','markdown','heading']:
        num_words = 0
        if 'source' in cell_keys:
            if isinstance(cell['source'], list):
                for l in cell['source']:
                    words = len(l.split())
                    num_words += words
            elif isinstance(cell['source'], str):
                num_words += len(cell['source'].split())
            else:
                num_words = None
    else:
        num_words = None
    
    if cell_type == 'code':
        lines_of_code = 0
        if 'source' in cell_keys:
            if isinstance(cell['source'], list):
                lines_of_code = len(cell['source'])
            elif isinstance(cell['source'], str):
                lines_of_code = len(cell['source'].splitlines())
            else:
                lines_of_code = None
            
        elif 'input' in cell_keys:
            if isinstance(cell['input'], list):
                lines_of_code = len(cell['input'])
            elif isinstance(cell['input'], str):
                lines_of_code = len(cell['input'].splitlines())
            else:
                lines_of_code = None
    else:
        lines_of_code = None
    
    
    # inilize output counts
    num_execute_result = 0
    num_error = 0
    num_stream = 0
    num_display_data = 0
    
    execute_result_keys = []
    error_names = []
    error_values = []
    display_data_keys = []
        
    if 'outputs' in cell_keys:
        for o in cell['outputs']:

            if isinstance(o, dict):
                output_keys = o.keys()
            else:
                output_keys = []

            if 'output_type' in output_keys:
                if o['output_type'] in ['execute_result','pyout']:
                    num_execute_result += 1
                    if 'data' in output_keys:
                        if isinstance(o['data'], dict):
                            data_keys = o['data'].keys()
                            for k in data_keys:
                                execute_result_keys.append(k)
                    else:
                        for k in output_keys:
                            if k in nbformat_3_mimes:
                                execute_result_keys.append(k)

                elif o['output_type'] in ['error','pyerr']:
                    num_error += 1
                    if 'ename' in output_keys:
                        error_names.append(o['ename'])
                    if 'evalue' in output_keys:
                        error_values.append(o['evalue'])

                elif o['output_type'] == 'stream':
                    num_stream += 1

                elif o['output_type'] == 'display_data':
                    num_display_data += 1
                    if 'data' in output_keys:
                        if isinstance(o['data'], dict):
                            data_keys = o['data'].keys()
                            for k in data_keys:
                                display_data_keys.append(k)
                    else:
                        for k in output_keys:
                            if k in nbformat_3_mimes:
                                display_data_keys.append(k)

    return [nb_id, worksheet_index, index, cell_type, num_words, lines_of_code, 
            num_execute_result, execute_result_keys, num_error, error_names, 
            error_values, num_stream, num_display_data, display_data_keys]


As is best practice, let's use a subset of the data to test our code

sample_cell_data = get_all_cell_data(df_nb_sample, 100)

And let's see how the celld ata looks

df_cell = pd.DataFrame(sample_cell_data)
print(df_cell.shape)
df_cell.columns = ['nb_id','workbook_index','cell_index','cell_type','num_words','lines_of_code',
                   'num_execute_result','execute_result_keys','num_error','error_names','error_values',
                   'num_stream','num_display_data','display_data_keys']
df_cell.head(20)

df_cell.to_csv('../data/csv/cell_metadata_sample.csv')

Alright, that computing of cell data seemed to work on 1000 notebooks, now for the 1.25 million

all_cell_data = get_all_cell_data(df_nb, 10000)

df_cell = pd.DataFrame(all_cell_data)
print(df_cell.shape)
df_cell.columns = ['nb_id','workbook_index','cell_index','cell_type','num_words','lines_of_code',
                   'num_execute_result','execute_result_keys','num_error','error_names','error_values',
                   'num_stream','num_display_data','display_data_keys']

df_cell.to_csv('../data/csv/cell_metadata.csv')

## Scrape nbformat 2.0 and 3.0 cells

Coming back from a break, it appears I did not check my code closely last time as I have a large number of files that did not get parsed. These are mainly the nbformat 2.0 and 3.0 files that store cells in a worksheet.

First up, I need a list of those files that did not make it into the list, igoring the ones that did not parse since these are simply due to malformatted json.

missing_cell_data = []

with open('../logs/cell_parse_log.txt', 'r') as f:
    for l in f:
        missing_cell_data.append(int(l.split(' ')[6]))

missing_cell_data = set(missing_cell_data)
missing_cell_data = np.array(list(missing_cell_data))
len(missing_cell_data)

df_cell = pd.read_csv('../data/csv/cell_metadata.csv')
nbs_with_cell_data = df_cell.nb_id.unique()

nbs_not_reparse = np.unique(np.append(missing_cell_data, nbs_with_cell_data))
len(nbs_not_reparse)

df_nbs_to_reparse = df_nb[~df_nb.nb_id.isin(nbs_not_reparse)]
df_nbs_to_reparse.shape

Okay, so we have ~173k notebooks that were not parsed and that are not a part of our known problem notebooks (i.e. ones with malformatted json that we could not get metadata from in step 1)

Let's test the revised code to make sure it works.

df_nbs_to_reparse_sample = df_nbs_to_reparse.sample(1000)

sample_reparse = get_all_cell_data(df_nbs_to_reparse_sample, 100)

df_sample_reparse = pd.DataFrame(sample_reparse)

df_sample_reparse.shape
df_sample_reparse.columns = ['nb_id','workbook_index','cell_index','cell_type','num_words','lines_of_code',
                   'num_execute_result','execute_result_keys','num_error','error_names','error_values',
                   'num_stream','num_display_data','display_data_keys']

len(df_sample_reparse['nb_id'].unique())

Yep, of the 1000 notebooks we tried to get data on, 13 didn't work. That is likely to more JSON malformatting issues. Okay, that sample worked, so now for all the missing data

reparse_cell_data = get_all_cell_data(df_nbs_to_reparse, 10000)

df_cell_reparse = pd.DataFrame(reparse_cell_data)
print(df_cell_reparse.shape)
df_cell_reparse.columns = ['nb_id','workbook_index','cell_index','cell_type','num_words','lines_of_code',
                   'num_execute_result','execute_result_keys','num_error','error_names','error_values',
                   'num_stream','num_display_data','display_data_keys']

df_cell_reparse.to_csv('../data/csv/cell_metadata_v2-3.csv')

reparsed_nbs = df_cell_reparse.nb_id.unique()

print(len(reparsed_nbs) + len(nbs_with_cell_data))
len(np.intersect1d(reparsed_nbs, nbs_not_reparse, assume_unique=True))

Between our first pass (which mainly got data from nbformat 4.0 notebooks) and this pass (which included 2.0 and 3.0 notebooks) we have 1.24 million notebooks, none of which overlap between the two lists. Now we can merge the two lists and save the updated csv file about cell data.

df_cell_all = pd.concat([df_cell_reparse, df_cell])
len(df_cell_all.nb_id.unique())

df_cell_all.to_csv('../data/csv/cell_metadata.csv')

## 3. Data about Cell Content

This will be the more complex parsing of data files. The main things we may want to get for now are imports, headers, and links. 

For the imports, I will need to watch for:
1. Python (import, from)
2. R (library, require)
3. Julia (using, import, importall)

I won't look at other languages as they all had less than 0.1 % of the full sample. 

For headers we can look for # in the markdown as well as the header cells in notebooks written in nbformat 2.0 - 3.0.

For links we may just use a regular expression. However, will this capture relative links? For that we may need to do another regular expression and check if the value returned looks like a url. If not, we may assume it is a relative link.

# Cell metadata
df_cell = pd.read_csv('../data/csv/cell_metadata.csv')
#df_nb.rename(columns = {'Unnamed: 0':'nb_id'}, inplace = True)
print(df_cell.shape)
df_cell.head()

df_nb_sample = df_nb.sample(1000)

def write_to_log(msg, log_name):
    f = '../logs/%s.txt' % log_name
    log_file = open(f, "a")
    log_file.write(msg + "\n")

def get_cell_content_data(df, tracking_trigger = 10000):
    
    all_cells = []      
    count = 0    
    time1 = time.time()
    
    for index, row in df.iterrows():
        
        # track progress through notebooks
        count += 1
        if count % tracking_trigger == 0:
            print('%s nbs in %s seconds' % (count, time.time() - time1))

        f = '../data/notebooks/nb_%s.ipynb' % row['nb_id']
        with open(f) as nb_file:
            date_string = datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S")

            try:
                data = json.load(nb_file)
            except:
                msg = '%s: had trouble loading nb %s' % (date_string, row['nb_id'])
                write_to_log(msg, 'cell_parse_content')
                continue

            if isinstance(data, dict): 
                keys = data.keys()
            else:
                keys = []
                
            # get the language
            nb_language = None
            if 'metadata' in keys:
                if isinstance(data, dict):
                    metadata_keys = data['metadata'].keys()
                else:
                    metadata_keys = []
            else:
                metadata_keys = []
            if 'language_info' in metadata_keys:
                if isinstance(data['metadata']['language_info'], dict):
                    lang_keys = data['metadata']['language_info'].keys()
                else:
                    lang_keys = None
                if 'name' in lang_keys:
                    nb_language = data['metadata']['language_info']['name']
            elif 'language' in metadata_keys:
                nb_language = data['metadata']['language']

            # get data for each cell, nbformat v 4.x
            if 'cells' in keys:
                for i, c in enumerate(data['cells']):
                    cell_data = get_cell_data(c, i, row['nb_id'], nb_language)
                    all_cells.append(cell_data)
            
            # get data for each cell, nbformat v 2.x / 3.x
            elif 'worksheets' in keys:
                for j, w in enumerate(data['worksheets']):
                    if isinstance(w, dict): 
                        worksheet_keys = w.keys()
                    else:
                        keys = []
                    if 'cells' in worksheet_keys:
                        for k, c in enumerate(w['cells']):
                            cell_data = get_cell_data(c, k, row['nb_id'], nb_language, j)
                            all_cells.append(cell_data)
                
    return all_cells
   

def get_cell_data(cell, index, nb_id, nb_language, worksheet_index = None):

    if isinstance(cell, dict): 
        cell_keys = cell.keys()
    else:
        cell_keys = [] 
    
    # get the cell type
    if 'cell_type' in cell_keys:
        cell_type = cell['cell_type']
    else:
        cell_type = None
    
    headings = []
    links = []
    imports = []
    
   # get the imports for python, Julia, and R
    if cell_type == 'code':
        
        lines_of_code = []
        
        if 'source' in cell_keys:
            if isinstance(cell['source'], list):
                lines_of_code = cell['source']
            elif isinstance(cell['source'], str):
                lines_of_code = cell['source'].splitlines()
            
        elif 'input' in cell_keys:
            if isinstance(cell['input'], list):
                lines_of_code = cell['input']
            elif isinstance(cell['input'], str):
                lines_of_code = cell['input'].splitlines()
                
        if nb_language == 'python':
            import_words = ['import', 'from']
        elif nb_language == 'R':
            import_words = ['library', 'require']
        elif nb_language in ['Julia', 'julia']:
            import_words = ['import', 'importall', 'using']
        else:
            import_words = []
            
        # get imports
        for l in lines_of_code:
            words = [x for x in re.split('\(|\)|\s*', l)if x]
            if len(words) >1:
                if words[0] in import_words:                    
                    imports.append(words[1])
    
    
    elif cell_type in ['heading', 'markdown']:
        
        
        # get the lines of markdown
        lines_of_markdown = []
        if 'source' in cell_keys:
            if isinstance(cell['source'], list):
                lines_of_markdown = cell['source']
            elif isinstance(cell['source'], str):
                lines_of_markdown = cell['source'].splitlines()
        
        # track headings in heading cells
        if 'level' in cell_keys:
            if 'source' in cell_keys:
                if len(lines_of_markdown) > 0:
                    headings.append([cell['level'], cell['source']])
        
        for l in lines_of_markdown:
            # get headings in markdown cells
            words = [x for x in re.split('\s*', l)if x]
            if len(words) > 2:
                if words[0] in ['#', '##', '###', '####', '#####', '######']:
                    headings.append([len(words[0]), l.split(words[0])[1]])


            # get links
            urls = re.findall('http[s]?://(?:[a-zA-Z]|[0-9]|[$-_@.&+]|[!*\(\),]|(?:%[0-9a-fA-F][0-9a-fA-F]))+', l)
            all_links = re.findall('\[([^]]+)]\(\s*(\S*)\)', l)
            for al in all_links:
                if al[:4] != 'http':
                    urls.append(al)
            for u in urls:
                links.append(u)
                
    return [nb_id, nb_language, index, worksheet_index, len(imports), imports, len(headings), headings, len(links), links]
        

all_cell_content = get_cell_content_data(df_nb)

df_cell_content = pd.DataFrame(all_cell_content)

df_cell_content.columns = ['nb_id', 'nb_language', 'cell_index', 'workbook_index', 
                           'num_imports', 'imports', 'num_headers', 'headers',
                           'num_links', 'links']

df_cell_meta_content = df_cell.merge(df_cell_content, how = 'left', on = ['nb_id', 'cell_index', 'workbook_index'])

del df_cell_meta_content['Unnamed: 0']
del df_cell_meta_content['Unnamed: 0.1']
del df_cell_meta_content['nb_language']

df_cell_meta_content.to_csv('../data/csv/cell_metadata_content.csv')

And now to the analysis! I think

## 4.0 More Data about Cell (code) Content 

After doing some of the later analysis for this project, we decided that we want a little more data about each notebook. Namely the code contents with comments, use of functions, and so on.

# Load cell metadata
df_cell = pd.read_csv('../data/csv/cell_metadata_content.csv')

print(df_cell.shape[0])
df_cell.head()

# get list of nbs to check
nbs_w_cells = df_cell['nb_id'].unique()
nbs_w_cells.shape[0]

df_nbs_cells = df_nb[df_nb['nb_id'].isin(nbs_w_cells)]

def write_to_log(msg, log_name):
    f = '../logs/%s.txt' % log_name
    log_file = open(f, "a")
    log_file.write(msg + "\n")

def get_cell_code_metadata(df, tracking_trigger = 10000):
    
    all_cells = []      
    count = 0    
    time1 = time.time()
    
    for index, row in df.iterrows():
        
        # track progress through notebooks
        count += 1
        if count % tracking_trigger == 0:
            print('%s nbs in %s seconds' % (count, time.time() - time1))

        f = '../data/notebooks/nb_%s.ipynb' % row['nb_id']
        with open(f) as nb_file:
            date_string = datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S")

            try:
                data = json.load(nb_file)
            except:
                msg = '%s: had trouble loading nb %s' % (date_string, row['nb_id'])
                write_to_log(msg, 'cell_parse_code_content')
                continue

            if isinstance(data, dict): 
                keys = data.keys()
            else:
                keys = []

            # get the language
            nb_language = None
            if 'metadata' in keys:
                if isinstance(data, dict):
                    metadata_keys = data['metadata'].keys()
                else:
                    metadata_keys = []
            else:
                metadata_keys = []
            if 'language_info' in metadata_keys:
                if isinstance(data['metadata']['language_info'], dict):
                    lang_keys = data['metadata']['language_info'].keys()
                else:
                    lang_keys = None
                if 'name' in lang_keys:
                    nb_language = data['metadata']['language_info']['name']
            elif 'language' in metadata_keys:
                nb_language = data['metadata']['language']
                
            # get data for each cell, nbformat v 4.x
            if 'cells' in keys:
                for i, c in enumerate(data['cells']):
                    cell_data = get_cell_data(c, i, row['nb_id'], nb_language)
                    all_cells.append(cell_data)
            
            # get data for each cell, nbformat v 2.x / 3.x
            elif 'worksheets' in keys:
                for j, w in enumerate(data['worksheets']):
                    if isinstance(w, dict): 
                        worksheet_keys = w.keys()
                    else:
                        keys = []
                    if 'cells' in worksheet_keys:
                        for k, c in enumerate(w['cells']):
                            cell_data = get_cell_data(c, k, row['nb_id'], nb_language, j)
                            all_cells.append(cell_data)
                
    return all_cells

def get_cell_data(cell, index, nb_id, nb_language, worksheet_index = None):

    if isinstance(cell, dict): 
        cell_keys = cell.keys()
    else:
        cell_keys = [] 
    
    # get the cell type
    if 'cell_type' in cell_keys:
        cell_type = cell['cell_type']
    else:
        cell_type = None
    
    functions = []
    classes = []
    comments = 0
    comments_words = 0
    in_multiline = False
    
   # get the imports for python, Julia, and R
    if cell_type == 'code':
        
        # get lines of code
        lines_of_code = []
        if 'source' in cell_keys:
            if isinstance(cell['source'], list):
                lines_of_code = cell['source']
            elif isinstance(cell['source'], str):
                lines_of_code = cell['source'].splitlines()          
        elif 'input' in cell_keys:
            if isinstance(cell['input'], list):
                lines_of_code = cell['input']
            elif isinstance(cell['input'], str):
                lines_of_code = cell['input'].splitlines()
                
        if nb_language == 'python':
            for l in lines_of_code:
                l = l.lstrip()
                parts = l.split()
                if len(parts) >= 2:                
                    if l.startswith('def'):
                        functions.append(parts[1].split('(')[0])
                    elif l.startswith('class'):
                        classes.append(parts[1].split('(')[0])
                
                if in_multiline:
                    comments += 1
                    comments_words += len(parts) - 1
                
                if l.startswith('"""'):
                    if not in_multiline:
                        comments += 1
                        comments_words += len(parts) - 1
                    in_multiline = not in_multiline
                elif l.startswith("'''"):
                    if not in_multiline:
                        comments += 1
                        comments_words += len(parts) - 1
                    in_multiline = not in_multiline
                elif l.startswith('#'):
                    comments += 1
                    comments_words += len(parts) - 1
                
                
    return [nb_id, nb_language, index, worksheet_index, len(functions), functions, len(classes), classes, comments, comments_words]

cell_code_content = get_cell_code_metadata(df_nbs_cells)

df_code = pd.DataFrame(cell_code_content)

df_code.columns = ['nb_id',
                    'language',
                    'cell_index',
                    'workbook_index',
                    'num_functions',
                    'name_functions',
                    'num_classes',
                    'name_classes',
                    'num_comment_lines',
                    'num_comment_words']
df_code.head()

df_code.to_csv('../data/csv/cell_metadata_code.csv')

And that's a wrap. Now we should have almost all the data we need to do our [core analysis of the notebooks](7_notebook_profiling.ipynb)
